Integrate cart and cart_items with TypeORM
