+++ src/app.module.ts
@@ -1,7 +1,7 @@
 import { Module } from '@nestjs/common';
 
 import { AppController } from './app.controller';
-import { Cart } from './entity'
+import { Cart, CartItem } from './entity'
 import { CartModule } from './cart/cart.module';
 import { AuthModule } from './auth/auth.module';
 import { OrderModule } from './order/order.module';
@@ -17,7 +17,7 @@ import { TypeOrmModule } from '@nestjs/typeorm';
     username: process.env.DB_USERNAME,
     password: process.env.DB_PASSWORD,
     database: process.env.DB_NAME,
-    entities: [Cart],
+    entities: [Cart,CartItem],
     synchronize: false,
     ssl: {
       rejectUnauthorized: false
