+++ src/entity/cart-status.ts
@@ -1,0 +1,4 @@
+export enum CartStatus {
+    OPEN = 'OPEN',
+    ORDERED = 'ORDERED'
+}
